Revert "Added helper function to rerender saved questions"
This reverts commit cfd07e1ad03ed8a2c399056e102d5be4c4ccf14b.
